+++ src/main/resources/db/migration/V1__Create_user_table.sql
@@ -1,7 +1,7 @@
 create table user
 (
-    id int auto_increment primary key not null,
-    account_id int,
+    id bigint auto_increment primary key not null,
+    account_id bigint,
     name varchar(50) default '',
     token varchar(36) default '',
     gmt_create bigint,
